Move action consts to actions.js
