--- wchat/redux/chats/constants.js
@@ -1,5 +1,0 @@
-export const ACTION = {
-  FETCH_CHATS : 'FETCH_CHATS',
-  FETCH_CHATS_SUCCESS: 'FETCH_CHATS_SUCCESS',
-  FETCH_CHATS_FAILURE : 'FETCH_CHATS_FAILURE',
-};
