--- wchat/redux/contacts/constants.js
@@ -1,5 +1,0 @@
-export const ACTION = {
-  FETCH_CONTACTS : 'FETCH_CONTACTS',
-  FETCH_CONTACTS_SUCCESS: 'FETCH_CONTACTS_SUCCESS',
-  FETCH_CONTACTS_FAILURE : 'FETCH_CONTACTS_FAILURE',
-};
