--- wchat/redux/groups/constants.js
@@ -1,5 +1,0 @@
-export const ACTION = {
-  FETCH_GROUPS : 'FETCH_GROUPS',
-  FETCH_GROUPS_SUCCESS: 'FETCH_GROUPS_SUCCESS',
-  FETCH_GROUPS_FAILURE : 'FETCH_GROUPS_FAILURE',
-};
